removed duplications
