+++ app/src/main/java/org/piramalswasthya/sakhi/adapters/FormInputAdapter.kt
@@ -5,7 +5,6 @@ import android.app.TimePickerDialog
 import android.content.Context
 import android.content.Context.INPUT_METHOD_SERVICE
 import android.content.res.ColorStateList
-import android.content.res.Configuration
 import android.content.res.Resources
 import android.graphics.Color
 import android.net.Uri
@@ -71,12 +70,12 @@ import org.piramalswasthya.sakhi.model.InputType.TIME_PICKER
 import org.piramalswasthya.sakhi.model.InputType.values
 import org.piramalswasthya.sakhi.ui.home_activity.all_ben.new_ben_registration.AgePickerDialog
 import org.piramalswasthya.sakhi.ui.home_activity.all_ben.new_ben_registration.ben_form.NewBenRegViewModel.Companion.isOtpVerified
+import org.piramalswasthya.sakhi.utils.HelperUtil
 import org.piramalswasthya.sakhi.utils.HelperUtil.findFragmentActivity
 import org.piramalswasthya.sakhi.utils.HelperUtil.getAgeStrFromAgeUnit
 import org.piramalswasthya.sakhi.utils.HelperUtil.getDobFromAge
 import org.piramalswasthya.sakhi.utils.HelperUtil.getLongFromDate
 import org.piramalswasthya.sakhi.utils.HelperUtil.updateAgeDTO
-import org.piramalswasthya.sakhi.utils.Log
 import timber.log.Timber
 import java.util.Calendar
 import java.util.Locale
@@ -698,13 +697,7 @@ class FormInputAdapter(
                 val activity = binding.et.context.findFragmentActivity()
                     ?: return@setOnClickListener
                 val originalLocale = Locale.getDefault()
-                Locale.setDefault(Locale.ENGLISH)
-                val config = Configuration(activity.resources.configuration)
-                config.setLocale(Locale.ENGLISH)
-                activity.resources.updateConfiguration(
-                    config,
-                    activity.resources.displayMetrics
-                )
+                HelperUtil.setEnLocaleForDatePicker(activity)
 
                 item.value?.let { value ->
                     thisYear = value.substring(6).toInt()
@@ -736,13 +729,7 @@ class FormInputAdapter(
                     datePickerDialog.datePicker.touchables[0].performClick()
                 datePickerDialog.show()
                 datePickerDialog.setOnDismissListener {
-                    Locale.setDefault(originalLocale)
-                    val restoreConfig = Configuration(activity.resources.configuration)
-                    restoreConfig.setLocale(originalLocale)
-                    activity.resources.updateConfiguration(
-                        restoreConfig,
-                        activity.resources.displayMetrics
-                    )
+                    HelperUtil.setOriginalLocaleForDatePicker(activity,originalLocale)
                 }
             }
             binding.executePendingBindings()
@@ -911,13 +898,7 @@ class FormInputAdapter(
                 val activity = binding.etDate.context.findFragmentActivity()
                     ?: return@setOnClickListener
                 val originalLocale = Locale.getDefault()
-                Locale.setDefault(Locale.ENGLISH)
-                val config = Configuration(activity.resources.configuration)
-                config.setLocale(Locale.ENGLISH)
-                activity.resources.updateConfiguration(
-                    config,
-                    activity.resources.displayMetrics
-                )
+                HelperUtil.setEnLocaleForDatePicker(activity)
 
                 item.value?.let { value ->
                     thisYear = value.substring(6).toInt()
@@ -953,13 +934,7 @@ class FormInputAdapter(
                     datePickerDialog.datePicker.touchables[0].performClick()
                 datePickerDialog.show()
                 datePickerDialog.setOnDismissListener {
-                    Locale.setDefault(originalLocale)
-                    val restoreConfig = Configuration(activity.resources.configuration)
-                    restoreConfig.setLocale(originalLocale)
-                    activity.resources.updateConfiguration(
-                        restoreConfig,
-                        activity.resources.displayMetrics
-                    )
+                    HelperUtil.setOriginalLocaleForDatePicker(activity,originalLocale)
                 }
             }
             binding.executePendingBindings()
