+++ app/src/main/java/org/piramalswasthya/sakhi/adapters/FormInputAdapterOld.kt
@@ -2,7 +2,6 @@ package org.piramalswasthya.sakhi.adapters
 
 import android.app.DatePickerDialog
 import android.app.TimePickerDialog
-import android.content.res.Configuration
 import android.net.Uri
 import android.os.CountDownTimer
 import android.text.Editable
@@ -56,8 +55,8 @@ import org.piramalswasthya.sakhi.model.InputType.RADIO
 import org.piramalswasthya.sakhi.model.InputType.TEXT_VIEW
 import org.piramalswasthya.sakhi.model.InputType.TIME_PICKER
 import org.piramalswasthya.sakhi.model.InputType.NUMBER_PICKER
-
 import org.piramalswasthya.sakhi.model.InputType.values
+import org.piramalswasthya.sakhi.utils.HelperUtil
 import org.piramalswasthya.sakhi.utils.HelperUtil.findFragmentActivity
 import timber.log.Timber
 import java.util.Calendar
@@ -514,13 +513,7 @@ class FormInputAdapterOld(
                 val activity = binding.et.context.findFragmentActivity()
                     ?: return@setOnClickListener
                 val originalLocale = Locale.getDefault()
-                Locale.setDefault(Locale.ENGLISH)
-                val config = Configuration(activity.resources.configuration)
-                config.setLocale(Locale.ENGLISH)
-                activity.resources.updateConfiguration(
-                    config,
-                    activity.resources.displayMetrics
-                )
+                HelperUtil.setEnLocaleForDatePicker(activity)
 
                 item.value.value?.let { value ->
                     thisYear = value.substring(6).toInt()
@@ -542,13 +535,7 @@ class FormInputAdapterOld(
                 datePickerDialog.datePicker.touchables[0].performClick()
                 datePickerDialog.show()
                 datePickerDialog.setOnDismissListener {
-                    Locale.setDefault(originalLocale)
-                    val restoreConfig = Configuration(activity.resources.configuration)
-                    restoreConfig.setLocale(originalLocale)
-                    activity.resources.updateConfiguration(
-                        restoreConfig,
-                        activity.resources.displayMetrics
-                    )
+                    HelperUtil.setOriginalLocaleForDatePicker(activity,originalLocale)
                 }
             }
             binding.executePendingBindings()
